feat: DetailReport show mapping code

+++ db/migrate/20140320100307_create_detail_reports.rb
@@ -4,7 +4,7 @@ class CreateDetailReports < ActiveRecord::Migration
       t.string :app_version
       t.string :test_datetime
       t.integer :status
-      t.integer :running_time
+      t.float :running_time
       t.string :device_key, index: true
       t.references :test_scenario, index: true
       t.references :total_report, index: true
